Move the UI support for DASD formatting to pyanaconda.ui.lib
Move the class DasdFormatting to pyanaconda.ui.lib.
